requirements.txt
# Core Telegram Bot Framework
python-telegram-bot==20.7

# Environment Variable Management
python-dotenv==1.0.0

# Task Scheduling (for future features like daily horoscopes)
APScheduler==3.10.4

# Additional useful packages for enhanced functionality
# Uncomment the ones you want to use:

# Date and time utilities (enhanced date handling)
# python-dateutil==2.8.2

# HTTP requests (for future API integrations)
# requests==2.31.0

# Data validation and parsing
# pydantic==2.5.0

# Timezone handling
# pytz==2023.3

# JSON handling and data processing
# ujson==5.8.0

# Performance monitoring (optional)
# psutil==5.9.6

# Testing dependencies (for development)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0

# Development tools (optional)
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1

# Production deployment (optional)
# gunicorn==21.2.0
# uvicorn==0.24.0

# Database utilities (if you want to switch from SQLite)
# sqlalchemy==2.0.23
# alembic==1.12.1

# Backup and data export (optional)
# pandas==2.1.3
# openpyxl==3.1.2